streamlit_app/requirements.txt
streamlit>=1.25.0
plotly>=5.14.0
altair>=5.0.0
pandas>=2.0.0
numpy>=1.24.0
networkx>=3.0
pillow>=9.5.0
requests>=2.28.0
python-dotenv>=0.20.0
jsonschema>=4.17.0
watchdog>=3.0.0
streamlit-extras>=0.3.0
openpyxl>=3.1.0 